chore: use session id for cache instead of token
