+++ db/people.go
@@ -3,6 +3,7 @@ package db
 import (
 	"github.com/flanksource/incident-commander/api"
 	"github.com/flanksource/incident-commander/utils"
+	"github.com/google/uuid"
 )
 
 func UpdateUserProperties(ctx *api.Context, userID string, newProps api.PersonProperties) error {
@@ -23,7 +24,13 @@ func UpdateIdentityState(ctx *api.Context, id, state string) error {
 	return ctx.DB().Table("identities").Where("id = ?", id).Update("state", state).Error
 }
 
-func CreateUser(ctx *api.Context, user api.Person) (api.Person, error) {
+func GetOrCreateUser(ctx *api.Context, user api.Person) (api.Person, error) {
+	if err := ctx.DB().Table("people").Where("email = ?", user.Email).Find(&user).Error; err != nil {
+		return api.Person{}, err
+	}
+	if user.ID != uuid.Nil {
+		return user, nil
+	}
 	err := ctx.DB().Table("people").Create(&user).Error
 	return user, err
 }
